our-game/game: refactoring
Add generic base class for spring controllers.

[#12]
