+++ game-server/application/spring-app/src/main/java/by/ourgame/game/application/spring/controller/SpringQuestionController.java
@@ -1,49 +1,21 @@
 package by.ourgame.game.application.spring.controller;
 
-import by.ourgame.game.adapter.controller.BaseController;
 import by.ourgame.game.domain.entity.Question;
 import by.ourgame.game.usecase.BaseDelete;
 import by.ourgame.game.usecase.BaseFind;
 import by.ourgame.game.usecase.BaseSave;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.*;
-
-import java.util.List;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 @RequestMapping("/questions")
-public class SpringQuestionController extends BaseController<Question> {
+public class SpringQuestionController extends BaseSpringController<Question> {
 
     @Autowired
     public SpringQuestionController(BaseSave<Question> saveQuestion,
                                     BaseFind<Question> findQuestion,
                                     BaseDelete<Question> deleteQuestion) {
         super(saveQuestion, findQuestion, deleteQuestion);
-    }
-
-    @PostMapping
-    @ResponseStatus(code = HttpStatus.CREATED)
-    @Override
-    public Question save(@RequestBody Question question) {
-        return super.save(question);
-    }
-
-    @GetMapping
-    @Override
-    public List<Question> findAll() {
-        return super.findAll();
-    }
-
-    @GetMapping("/{id}")
-    @Override
-    public Question findById(@PathVariable String id) {
-        return super.findById(id);
-    }
-
-    @DeleteMapping("/{id}")
-    @Override
-    public void deleteById(@PathVariable String id) {
-        super.deleteById(id);
     }
 }
